Renamed request_info component files and deleted signup.css file.
