+++ frontend/web/src/components/home/body/dashboard/user_dash/request_info/request_info.js
@@ -1,7 +1,7 @@
 import React, { Fragment, useContext, useEffect, useState } from "react";
 import { useHistory } from "react-router-dom";
 import { RequestContext } from "../../../../../utility/contexts/requestContext";
-import "./requestInfo.css";
+import "./request_info.css";
 import axios from "axios";
 import { Button, Comment, Form, Header } from "semantic-ui-react";
 import Avatar from "../../../../../../asserts/images/avatar.png";
